fix:skip inclusion node_modules form current_cloud_backend.zip
Current cloud backend included Lambda functions node_modules. This can cause errors when checking
out an env where node_modules has reference to local node_package. By skipping node_modules folder
the size of current_cloud_backend.zip file size reduces and also fixes the error when referencing
local packages

fixes #1077
